Use R8snorm instead or R8uint for marker Seems to be slightly faster than R8uint, but more importantly it paves the way for curved boundaries

+++ shader/volume_visualization/velocity.vert
@@ -27,7 +27,7 @@ void main() {
     ivec3 volumeCoordinate = getVolumeCoordinate(gl_InstanceIndex / 3);
     uint channel = gl_InstanceIndex % 3;
 
-    uint marker = texelFetch(MarkerVolume, volumeCoordinate, 0).x;
+    float marker = texelFetch(MarkerVolume, volumeCoordinate, 0).x;
 
     vec3 cellCenter = (volumeCoordinate + vec3(0.5)) * Rendering.FluidGridToWorldScale + Rendering.FluidWorldOrigin;
     vec3 linePosition = cellCenter;
